# sample-api-documentation.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Megaexe Post Management",
      "description": "Megaexe project task",
      "version": "1.0",
      "contact": {
  
      }
    },
...
    "servers": [],
    "components": {
      "securitySchemes": {
        "bearer": {
          "scheme": "bearer",
          "bearerFormat": "JWT",
          "type": "http"
        }
      },
      "schemas": {
...
        "UpdatePostDto": {
          "type": "object",
          "properties": {
            "post_category_id": {
              "type": "string"
            },
            "title": {
              "type": "string"
            },
            "content": {
              "type": "string"
            },
            "image": {
              "type": "string"
            }
          }
        },
...
        "CreateCommentDto": {
          "type": "object",
          "properties": {
            "post_id": {
              "type": "string"
            },
            "comment": {
              "type": "string"
            }
          },
          "required": [
            "post_id",
            "comment"
          ]
        },
        "UpdateCommentDto": {
          "type": "object",
          "properties": {
            "post_id": {
              "type": "string"
            },
            "comment": {
              "type": "string"
            }
          }
        }
      }
    },
    "paths": {
...
      "/api/v1/post-category/active": {
        "get": {
          "operationId": "findAllActive",
          "parameters": [],
          "responses": {
            "200": {
              "description": ""
            }
          },
          "tags": [
            "Post Category"
          ],
          "security": [
            {
              "bearer": []
            }
          ]
        }
      },
...
      "/api/v1/post/category/{category_id}": {
        "get": {
          "operationId": "findAllActive",
          "parameters": [
            {
              "name": "category_id",
              "required": true,
              "in": "path",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": ""
            }
          },
          "tags": [
            "Post Management"
          ],
          "security": [
            {
              "bearer": []
            }
          ]
        }
      },
      "/api/v1/post/{id}": {
        "get": {
          "operationId": "findOne",
          "parameters": [
            {
              "name": "id",
              "required": true,
              "in": "path",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": ""
            }
          },
          "tags": [
            "Post Management"
          ],
          "security": [
            {
              "bearer": []
            }
          ]
        },
        "patch": {
          "operationId": "update",
          "parameters": [
            {
              "name": "id",
              "required": true,
              "in": "path",
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatePostDto"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": ""
            }
          },
          "tags": [
            "Post Management"
          ],
          "security": [
            {
              "bearer": []
            }
          ]
        },
        "delete": {
          "operationId": "remove",
          "parameters": [
            {
              "name": "id",
              "required": true,
              "in": "path",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": ""
            }
          },
          "tags": [
            "Post Management"
          ],
          "security": [
            {
              "bearer": []
            }
          ]
        }
      },
...
      "/api/v1/comments": {
        "post": {
          "operationId": "create",
          "parameters": [],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateCommentDto"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": ""
            }
          },
          "tags": [
            "Comment Management"
          ],
          "security": [
            {
              "bearer": []
            }
          ]
        },
        "get": {
          "operationId": "findAll",
          "parameters": [
            {
              "name": "page",
              "required": false,
              "in": "query",
              "schema": {
                "type": "number"
              }
            },
            {
              "name": "per_page",
              "required": false,
              "in": "query",
              "schema": {
                "type": "number"
              }
            },
            {
              "name": "post_id",
              "required": false,
              "in": "query",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "search",
              "required": false,
              "in": "query",
              "schema": {
                "type": "string"
              }
            },
            {
              "name": "user_id",
              "required": false,
              "in": "query",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": ""
            }
          },
          "tags": [
            "Comment Management"
          ],
          "security": [
            {
              "bearer": []
            }
          ]
        }
      },
      "/api/v1/comments/post/{post_id}": {
        "get": {
          "operationId": "findAllActive",
          "parameters": [
            {
              "name": "post_id",
              "required": true,
              "in": "path",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": ""
            }
          },
          "tags": [
            "Comment Management"
          ],
          "security": [
            {
              "bearer": []
            }
          ]
        }
      },
      "/api/v1/comments/{id}": {
        "get": {
          "operationId": "findOne",
          "parameters": [
            {
              "name": "id",
              "required": true,
              "in": "path",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": ""
            }
          },
          "tags": [
            "Comment Management"
          ],
          "security": [
            {
              "bearer": []
            }
          ]
        },
        "patch": {
          "operationId": "update",
          "parameters": [
            {
              "name": "id",
              "required": true,
              "in": "path",
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateCommentDto"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": ""
            }
          },
          "tags": [
            "Comment Management"
          ],
          "security": [
            {
              "bearer": []
            }
          ]
        },
        "delete": {
          "operationId": "remove",
          "parameters": [
            {
              "name": "id",
              "required": true,
              "in": "path",
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": ""
            }
          },
          "tags": [
            "Comment Management"
          ],
          "security": [
            {
              "bearer": []
            }
          ]
        }
      }
    }
  }